feat(viewer): add textMove mode and text move util function
