+++ packages/insight-viewer/src/utils/common/getEditingTextPosition.ts
@@ -1,0 +1,5 @@
+import { Point } from '../../types'
+
+export function getEditingTextPosition(currentPoint: Point): Point {
+  return currentPoint
+}
