+++ packages/insight-viewer/src/utils/common/getTextPosition.ts
@@ -1,19 +1,18 @@
-import { getRulerTextPosition } from './getRulerTextPosition'
-import { getCircleTextPosition } from './getCircleTextPosition'
-import { getLineLengthWithoutImage } from './getLineLengthWithoutImage'
+import { getDrawingTextPosition } from './getDrawingTextPosition'
+import { getEditingTextPosition } from './getEditingTextPosition'
 
-import { Point, MeasurementMode } from '../../types'
+import { Point, EditMode, Measurement, MeasurementMode } from '../../types'
 
-export function getTextPosition(points: [Point, Point], mode: MeasurementMode): Point {
-  if (mode === 'ruler') {
-    // index 1 of Points array is the end point in ruler mode
-    const rulerTextPosition = getRulerTextPosition(points[1])
-
-    return rulerTextPosition
+export function getTextPosition(
+  mode: MeasurementMode,
+  points: [Point, Point],
+  measurement: Measurement | null,
+  editMode?: EditMode | null,
+  currentPoint?: Point
+): Point {
+  if (editMode === 'textMove' && currentPoint) {
+    return getEditingTextPosition(currentPoint)
   }
-  // mode === 'circle'
-  const drawingRadius = getLineLengthWithoutImage(points[0], points[1])
-  const circleTextPosition = getCircleTextPosition(points[0], drawingRadius)
 
-  return circleTextPosition
+  return getDrawingTextPosition(points, mode, measurement)
 }
